+++ playwright/ci-test/tests/project/create_complex.ts
@@ -111,8 +111,8 @@ test.describe('Create complex project', () => {
     await page.getByRole('option', { name: 'Sample Indicator A' }).click();
     await page.getByRole('combobox').nth(3).click();
     await page.getByRole('option', { name: 'value' }).click();
-    await page.getByRole('button', { name: 'Apply' }).click();
     await expect(page.locator('label').filter({ hasText: 'No filter (global latest' })).toBeDisabled();
+    await page.getByRole('button', { name: 'Apply' }).click();
 
     // Update tools
     await page.locator('.TabPrimary').getByText('Tools').click();
